DocRecServer/requirements.txt
aiodns==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.1
amqp==5.1.1
anyio==3.5.0
APScheduler==3.10.1
asgiref==3.6.0
async-timeout==4.0.2
asyncpg==0.25.0
attrs==23.1.0
billiard==3.6.4.0
ccxt==4.0.39
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==41.0.2
dateparser==1.1.8
fastapi==0.75.0
frozenlist==1.4.0
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
passlib==1.7.4
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pybit==5.2.0
pycares==4.3.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.9.0
PyJWT==2.3.0
python-binance==1.0.17
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
redis==4.5.4
regex==2023.6.3
requests==2.30.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
typing_extensions==4.1.1
tzdata==2023.3
tzlocal==4.3
ujson==5.8.0
urllib3==2.0.2
uvicorn==0.17.6
uvloop==0.17.0
vine==5.0.0
watchfiles==0.18.1
wcwidth==0.2.6
websocket-client==1.5.1
websockets==10.1
yarl==1.9.2
